+++ richat/src/bin/richat.rs
@@ -18,6 +18,10 @@ use {
     tracing::{info, warn},
 };
 
+#[cfg(not(target_env = "msvc"))]
+#[global_allocator]
+static GLOBAL: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
+
 #[derive(Debug, Parser)]
 #[clap(author, version, about = "Richat App")]
 struct Args {